Module cargo_auto::inside_of_rust_project_mod

Functions

already_exists_automation_tasks_rs 🔒

already exists automation_tasks_rs directory

completion 🔒

sub-command for bash auto-completion of cargo auto using the crate dev_bestia_cargo_completion

match_first_argument 🔒

the first argument is the task: new_auto, build, release,…
the task new is processed by cargo-auto,
all other tasks are processed by the used defined automation_tasks_rs
in development use: cargo run -- new_auto
in development use: cargo run -- build
in development use: cargo run -- release

parse_args

print_help_from_cargo_auto 🔒

if there is no argument then print help
if there exists automation_tasks_rs/Cargo.toml and automation_tasks_rs/src/main.rs
call automation_tasks_rs with no arguments to print the help prepared in user defined automation_tasks_rs
else print the help for cargo auto new_auto
in development use: cargo run
in runtime use: cargo auto